% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_panel.r
\name{balance_panel}
\alias{balance_panel}
\title{balance a time series panel}
\usage{
balance_panel(df, unit = unit, time = time, start = NULL,
  stop = NULL, complete_obs = TRUE)
}
\arguments{
\item{df}{a tibble}

\item{unit}{(optional) the unit variable, default is unit}

\item{time}{(optional) the time variable, default is time}

\item{start}{(optional) the first year in the panel (int), default is
minimum observed}

\item{stop}{(optional) the last year in the panel (int), default is the
maximum observed}

\item{complete_obs}{(optional) require units to be observed in all time
periods? default is true.}
}
\value{
input tibble filtered to balanced panel
}
\description{
Balances a panel of unit-time data. Filters out units that are
missing observations.
}
\details{
Give a panel of observations, this function subsets the data frame
to only include those units that are observed in every time period.
Optional `start` and `stop` times can be designated, if specified,
observations before the start and stop times will be dropped.
Set `complete_obs = FALSE` to keep units that are missing some time
observations during the panel.
}
\examples{
 balance_panel(df = airquality, unit = Month, time = Day)

}
