% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vars_exist.R
\name{check_vars_exist}
\alias{check_vars_exist}
\title{check variable existence in data set}
\usage{
check_vars_exist(df, ...)
}
\arguments{
\item{df}{a tibble}

\item{...}{list of variables to confirm existence of.}
}
\value{
Silent or Error
}
\description{
given a list of variables and a data set, checks if those variables
exist and throws an error stating 'variable x not found' for whichever ones
don't. I got fed up with opaque error messages caused by mistyped variable
names. If you are nesting this inside another function and using tidy style
remember to unquote (!!) the variables that you pass in. The function returns
nothing if the variables exist.
}
\examples{
check_vars_exist(df = airquality, Ozone, Month)
}
