% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_sample.R
\name{g_sample}
\alias{g_sample}
\alias{g_sample_frac}
\alias{g_sample_n}
\title{Sample groups from a table}
\usage{
g_sample_frac(tbl, group, ...)

g_sample_n(tbl, group, ...)
}
\arguments{
\item{tbl}{tbl of data}

\item{group}{grouping variable}

\item{...}{Arguments passed on to \code{dplyr::sample_frac}
\describe{
  \item{size}{For \code{sample_n()}, the number of rows to select.
For \code{sample_frac()}, the fraction of rows to select.
If \code{tbl} is grouped, \code{size} applies to each group.}
  \item{replace}{Sample with or without replacement?}
  \item{weight}{Sampling weights. This must evaluate to a vector of
non-negative numbers the same length as the input. Weights are
automatically standardised to sum to 1.

This argument is automatically \link[rlang:quo]{quoted} and later
\link[rlang:eval_tidy]{evaluated} in the context of the data
frame. It supports \link[rlang:quasiquotation]{unquoting}. See
\code{vignette("programming")} for an introduction to these concepts.}
}}
}
\description{
using group_by with sample returns a sample of the observations
within a group. This function instead returns all observations for a sample
of groups.
}
\examples{
g_sample_frac(airquality, Month, .3)
g_sample_n(airquality, Month, 3)
}
\seealso{
See \code{\link[dplyr]{sample}} for details on sample function.
}
