% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liefTools-package.R
\docType{package}
\name{liefTools-package}
\alias{liefTools}
\alias{liefTools-package}
\title{liefTools: Lief's personal functions}
\description{
Various and sundry helper functions that I use.
}
\keyword{internal}
