% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liefTools.r
\docType{package}
\name{liefTools}
\alias{liefTools}
\alias{liefTools-package}
\title{liefTools package}
\description{
My personal useful functions
}
\details{
See the readme on \href{https://github.com/williamlief/liefTools}{Github}
}
