% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_all_sheets.R
\name{read_excel_all}
\alias{read_excel_all}
\title{Read all sheets in an excel file}
\usage{
read_excel_all(path, ..., sheet_id = ".sheet")
}
\arguments{
\item{path}{Path to the xls/xlsx file}

\item{...}{all arguments to readxl::read_excel() EXCEPT sheet. If you try to
explicitly use a sheet param the function will break.}

\item{sheet_id}{default = ".sheet": name of column that contains the sheet_name}
}
\value{
a tibble
}
\description{
This function pulls in all sheets in an excel document. It
  passes the same arguments for every sheet. Thefore it is best used on excel
  files where the sheets have identical layouts but have split the data into
  separate sheets by some grouping characteristic (e.g. state, year,
  company). An additional column (name controlled by `sheet_id` paramater)
  tracks the name of each sheet in the returned tibble.

  It takes all the parameters of read_excel EXCEPT 'sheet', the list of
  sheets is obtained through a call to `readxl::excel_sheets()`. Future
  versions of this function should include a `sheets` parameter to allow the
  user to specify explicitly which sheets to read.

  Stackoverflow for help building the function:
  \code{\link{https://stackoverflow.com/q/58243710/3373438}}
}
\examples{
read_excel_all(readxl::readxl_example("datasets.xlsx"), n_max = 2)
}
\seealso{
\code{\link[readxl]{read_excel}}
}
