% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_frac_g.R
\name{sample_frac_g}
\alias{sample_frac_g}
\title{Sample fraction of groups}
\usage{
sample_frac_g(tbl, group, size = 1, replace = FALSE, weight = NULL)
}
\arguments{
\item{tbl}{tbl of data}

\item{group}{grouping variable}

\item{size}{the fraction of groups to select}

\item{replace}{sample with or without replacement}

\item{weight}{sampling weights // NOT SUPPORTED}
}
\description{
using group_by with sample_frac returns a sample of the observations
within a group. This function instead returns all observations for a fraction
of groups.
}
\examples{
sample_frac_g(airquality, Month, .3)
}
