% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ext.R
\name{summarize_mean}
\alias{summarize_mean}
\title{Summary of mean by variable}
\usage{
summarize_mean(df, group = NULL, round = NULL, na.rm = TRUE)
}
\arguments{
\item{df}{a tibble}

\item{group}{(optional) single grouping variable. If used, mean by
group will be spread across columns.}

\item{round}{(optional) specify number of digits to round means to. Default is
no rounding.}

\item{na.rm}{(optional) a logical value indicating whether NA values should be
stripped before the computation proceeds. Unlike base, here I default to TRUE}
}
\value{
a tibble with one row per variable in original datset, values are mean
}
\description{
reports mean value for all numeric and logical variables in a
data frame
}
\examples{
 summarize_mean(df = airquality)
 summarize_mean(df = airquality, group = Month)
 summarize_mean(df = airquality, group = Month, round = 3)
 summarize_mean(df = airquality, na.rm = FALSE)
}
