% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ext.R
\name{summarize_stats}
\alias{summarize_stats}
\title{mean, sd, min, max of variables}
\usage{
summarize_stats(tbl, ..., na.rm = TRUE)
}
\arguments{
\item{tbl}{A tbl}

\item{...}{variables to summarize}

\item{na.rm}{logical value indicating whether NA values should be stripped
before the computation proceeds. Default is TRUE}
}
\value{
a long table with one row per variable. Columns are variable, mean, sd,
min and max.
}
\description{
returns the mean, sd, min and max of a list of variables. By default
passes na.rm = TRUE. Respects grouping of input data.
}
\details{
This function respects grouping. If the tbl is grouped, the output
will be as well.
}
\examples{
summarize_stats(airquality, Ozone, Solar.R, Wind, Temp)
summarize_stats(airquality \%>\% dplyr::group_by(Month), Ozone, Solar.R, Wind, Temp)
}
