% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_sample.R
\name{g_sample}
\alias{g_sample}
\alias{g_sample_frac}
\alias{g_sample_n}
\title{Sample groups from a table}
\usage{
g_sample_frac(tbl, group, ...)

g_sample_n(tbl, group, ...)
}
\arguments{
\item{tbl}{tbl of data}

\item{group}{grouping variable}

\item{...}{ignored}
}
\description{
using group_by with sample returns a sample of the observations
within a group. This function instead returns all observations for a sample
of groups.
}
\examples{
g_sample_frac(airquality, Month, .3)
g_sample_n(airquality, Month, 3)
}
\seealso{
\code{\link[dplyr]{sample}}
}
