% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_na.R
\name{summarize_na}
\alias{summarize_na}
\title{Summary of NA by variable}
\usage{
summarize_na(df, group = NULL)
}
\arguments{
\item{df}{: a tibble}

\item{group}{: optional single grouping variable. If used, pct missing by
group will be spread across columns.}
}
\value{
a tibble with one row per variable in original datset, values are pct na.
}
\description{
reports percent of missing data for all variables in data frame
}
\examples{
 summarize_na(df = airquality)
 summarize_na(df = airquality, group = Month)
}
